requirements.txt
slackclient
flask
bottle